#  feat(query): Parallelize scanPartitions and all QE work in query threadpool