--- a/core/src/main/resources/filodb-defaults.conf
+++ b/core/src/main/resources/filodb-defaults.conf
@@ -38,6 +38,9 @@ filodb {
 
     # Minimum step required for a query
     min-step = 5 seconds
+
+    # Number of threads in query threadpool.  There is one threadpool per dataset.
+    num-threads = 8
   }
 
   shard-manager {