--- a/core/src/main/resources/filodb-defaults.conf
+++ b/core/src/main/resources/filodb-defaults.conf
@@ -39,8 +39,8 @@ filodb {
     # Minimum step required for a query
     min-step = 5 seconds
 
-    # Number of threads in query threadpool.  There is one threadpool per dataset.
-    num-threads = 8
+    # Parallelism (query threadpool per dataset) ... ceil(available processors * factor)
+    threads-factor = 1.0
   }
 
   shard-manager {