**Pull Request checklist**

- [x] The commit(s) message(s) follows the contribution [guidelines](CONTRIBUTING.md) ?
- [x] Tests for the changes have been added (for bug fixes / features) ?
- [ ] Docs have been added / updated (for bug fixes / features) ?

**Current behavior :** (link exiting issues here : https://help.github.com/articles/basic-writing-and-formatting-syntax/#referencing-issues-and-pull-requests)

The setup of each query, including `scanPartitions`, Lucene partId lookup, and setting up RangeVectorTransformers, happens single-threaded in the Actor receive loop's threadpool.   The overhead of this setup being run in serial can be significant.

**New behavior :**

1.  The `ExecPlan.execute` method is run entirely inside of a Monix Task, including all setup, `scanPartitions`, etc.  This means these tasks can now be parallelized and execute outside of the Actor loop.
2. All of the QueryEngine tasks and Observables are now run in a separate, configurable threadpool, named query-<dataset>.  This will make it easier to do profiling later too.

### Before

```
[info] Benchmark                                               Mode  Cnt      Score     Error  Units
[info] QueryInMemoryBenchmark.noOverlapQueries                thrpt   40   3724.422 ±  94.726  ops/s
[info] QueryInMemoryBenchmark.singleThreadedMinOverTimeQuery  thrpt   40  13859.802 ± 197.079  ops/s
[info] QueryInMemoryBenchmark.singleThreadedRawQuery          thrpt   40  24375.510 ± 385.174  ops/s
[info] QueryInMemoryBenchmark.singleThreadedSumRateCCQuery    thrpt   40   4626.947 ±  57.162  ops/s
[info] QueryInMemoryBenchmark.someOverlapQueries              thrpt   40   3112.174 ±  88.544  ops/s
```

### After

```
[info] Benchmark                                               Mode  Cnt      Score     Error  Units
[info] QueryInMemoryBenchmark.noOverlapQueries                thrpt   40   9521.884 ± 142.717  ops/s
[info] QueryInMemoryBenchmark.singleThreadedMinOverTimeQuery  thrpt   40  14027.576 ± 128.490  ops/s
[info] QueryInMemoryBenchmark.singleThreadedRawQuery          thrpt   40  23583.630 ± 553.107  ops/s
[info] QueryInMemoryBenchmark.singleThreadedSumRateCCQuery    thrpt   40   4627.667 ±  37.934  ops/s
[info] QueryInMemoryBenchmark.someOverlapQueries              thrpt   40   5756.704 ± 221.029  ops/s
```