@vishramachandran 2019-03-25
@velvia  @broneill  A thought for the future.
I don't see a need for all queries to be routed through a single actor (QueryActor). There is no need for serialization at all whatsoever here.
We could simply expose a grpc based HTTP api and route queries directly from there.